slurm resources moved to slurm namespace
