+++ pegasus_wrapper/scripts/nuke_checkpoints.py
